+++ src/context/directory/handlers/actions.js
@@ -14,7 +14,9 @@ function parse(context) {
   const actions = files.map((file) => {
     const action = { ...loadJSON(file, context.mappings) };
     const actionFolder = path.join(constants.ACTIONS_DIRECTORY, `${action.name}`);
-    action.code = context.loadFile(action.code, actionFolder);
+    if(action.code){
+      action.code = context.loadFile(action.code, actionFolder);
+    }
     if (action.current_version && JSON.stringify(action.current_version) !== JSON.stringify({})) {
       action.current_version.code = context.loadFile(action.current_version.code, constants.ACTIONS_DIRECTORY);
     }
@@ -23,6 +25,13 @@ function parse(context) {
   return {
     actions
   };
+}
+
+function mapSecrets(secrets){
+  if(secrets && secrets.length > 0){
+    return secrets.map(secret => ({name:secret.name, value: secret.value}))
+  }
+  return [];
 }
 
 function mapCurrentVersion(filePath, action) {
@@ -45,10 +54,8 @@ function mapCurrentVersion(filePath, action) {
     code: `${codeFile}`,
     number: version.number,
     dependencies: version.dependencies || [],
-    secrets: version.secrets || [],
+    secrets: mapSecrets(version.secrets),
     runtime: version.runtime,
-    created_at: version.created_at,
-    updated_at: version.updated_at
   };
 }
 
@@ -76,7 +83,7 @@ function mapToAction(filePath, action) {
     code: mapActionCode(filePath, action),
     status: action.status,
     dependencies: action.dependencies || [],
-    secrets: action.secrets || [],
+    secrets: mapSecrets(action.secrets),
     runtime: action.runtime,
     supported_triggers: action.supported_triggers,
     current_version: mapCurrentVersion(filePath, action)
@@ -93,7 +100,7 @@ async function dump(context) {
   actions.forEach((action) => {
     // Dump template metadata
     const name = sanitize(action.name);
-    const actionFile = path.join(actionsFolder, name, `${name}.json`);
+    const actionFile = path.join(actionsFolder, `${name}.json`);
     log.info(`Writing ${actionFile}`);
     fs.writeFileSync(actionFile, JSON.stringify(mapToAction(context.filePath, action), null, 2));
   });